smart协议的功能都可以了;rxt增加信号量分配接口

+++ inc/os/rtx/rtx_depend.h
@@ -13,7 +13,7 @@
 //#include "osObjects.h"                      // RTOS object definitions
 typedef unsigned int os_time_t;
 typedef int mutext_t;
-typedef uint8_t sem_t;
+typedef int sem_t;
 
 
 #endif /* OS_DEPEND_H_ */
